+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### [28.07.26 - 0.0.4]
+- added GitlabCiBuilderImpl
+- added BaseLinux and _base param for JobBuilderImpl
+- changed base tags from `immortal` to `static-k8s`
+
 ### [31.05.26 - 0.0.3]
 - removed pipeline annotation
 - fixed workflow for gitlab-ci and pipeline